+++ src/Model/Reports/HourReportFormData.php
@@ -3,12 +3,14 @@
 namespace App\Model\Reports;
 
 use App\Entity\DataProvider;
+use App\Entity\Project;
+use App\Entity\Version;
 
 class HourReportFormData
 {
     public DataProvider $dataProvider;
-    public string $projectId;
-    public string $versionId;
+    public Project $project;
+    public Version $version;
     public \DateTimeInterface $fromDate;
     public \DateTimeInterface $toDate;
 }