+++ src/Entity/DataProvider.php
@@ -3,8 +3,6 @@
 namespace App\Entity;
 
 use App\Repository\DataProviderRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 
 #[ORM\Entity(repositoryClass: DataProviderRepository::class)]
@@ -28,16 +26,17 @@ class DataProvider extends AbstractBaseEntity
     #[ORM\Column(nullable: true)]
     private ?bool $enableAccountSync = null;
 
-    #[ORM\ManyToMany(targetEntity: View::class, mappedBy: 'dataProviders')]
-    private Collection $views;
+    #[ORM\Column(nullable: true)]
+    private ?bool $enabled = null;
 
     public function __construct()
     {
-        $this->views = new ArrayCollection();
     }
 
-    #[ORM\Column(nullable: true)]
-    private ?bool $enabled = null;
+    public function __toString(): string
+    {
+        return $this->getName() ?? ''.$this->getId();
+    }
 
     public function getName(): ?string
     {
@@ -107,38 +106,6 @@ class DataProvider extends AbstractBaseEntity
     public function setEnableAccountSync(bool $enableAccountSync): static
     {
         $this->enableAccountSync = $enableAccountSync;
-
-        return $this;
-    }
-
-    public function __toString(): string
-    {
-        return $this->getName() ?? ''.$this->getId();
-    }
-
-    /**
-     * @return Collection<int, View>
-     */
-    public function getViews(): Collection
-    {
-        return $this->views;
-    }
-
-    public function addView(View $view): static
-    {
-        if (!$this->views->contains($view)) {
-            $this->views->add($view);
-            $view->addDataProvider($this);
-        }
-
-        return $this;
-    }
-
-    public function removeView(View $view): static
-    {
-        if ($this->views->removeElement($view)) {
-            $view->removeDataProvider($this);
-        }
 
         return $this;
     }