+++ src/Service/DataSynchronizationService.php
@@ -314,7 +314,15 @@ class DataSynchronizationService
                 throw new EconomicsException($this->translator->trans('exception.project_not_found'));
             }
 
+            $issue = !empty($worklogDatum->projectTrackerIssueId) ? $this->issueRepository->findOneBy(['projectTrackerId' => $worklogDatum->projectTrackerIssueId]) : null;
+
+            if ($issue === null) {
+                // A worklog should always have an issue, so ignore the worklog.
+                continue;
+            }
+
             $worklog = $this->worklogRepository->findOneBy(['worklogId' => $worklogDatum->projectTrackerId]);
+
             if (!$worklog) {
                 $worklog = new Worklog();
 
@@ -324,18 +332,15 @@ class DataSynchronizationService
 
                 $this->entityManager->persist($worklog);
             }
+
             $worklog
                 ->setWorklogId($worklogDatum->projectTrackerId)
                 ->setDescription($worklogDatum->comment)
                 ->setWorker($worklogDatum->worker)
                 ->setStarted($worklogDatum->started)
                 ->setProjectTrackerIssueId($worklogDatum->projectTrackerIssueId)
-                ->setTimeSpentSeconds($worklogDatum->timeSpentSeconds);
-
-            if (null != $worklog->getProjectTrackerIssueId()) {
-                $issue = $this->issueRepository->findOneBy(['projectTrackerId' => $worklog->getProjectTrackerIssueId()]);
-                $worklog->setIssue($issue);
-            }
+                ->setTimeSpentSeconds($worklogDatum->timeSpentSeconds)
+                ->setIssue($issue);
 
             if (!$worklog->isBilled() && $worklogDatum->projectTrackerIsBilled) {
                 $worklog->setIsBilled(true);