+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* [PR-135](https://github.com/itk-dev/economics/pull/135)
+  1772: Removed views.
 * [PR-134](https://github.com/itk-dev/economics/pull/134)
   1632: Remove team report.
 * [PR-133](https://github.com/itk-dev/economics/pull/133)