+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [PR-135](https://github.com/itk-dev/economics/pull/135)
   1772: Removed views.
+* [PR-137](https://github.com/itk-dev/economics/pull/137)
+  1812: Minor hour report improvements.
 * [PR-134](https://github.com/itk-dev/economics/pull/134)
   1632: Remove team report.
 * [PR-133](https://github.com/itk-dev/economics/pull/133)