+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Added check that issue exists before adding worklog to database.
+* [PR-139](https://github.com/itk-dev/economics/pull/139)
+  1890: Added check that issue exists before adding worklog to database.
 
 ## [2.3.0] - 2024-06-03