+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.1] - 2024-07-05
+
 * [PR-139](https://github.com/itk-dev/economics/pull/139)
   1890: Added check that issue exists before adding worklog to database.